topchef_app/requirements.txt
requests
psycopg2-binary
python-dotenv
schedule
openai
fastapi
uvicorn[standard]
jinja2
